SCAN4NET-335 Extract ProxyTest
