+++ its/src/test/java/com/sonar/it/scanner/msbuild/sonarqube/ProxyTest.java
@@ -1,0 +1,212 @@
+/*
+ * SonarScanner for .NET
+ * Copyright (C) 2016-2025 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.sonar.it.scanner.msbuild.sonarqube;
+
+import com.sonar.it.scanner.msbuild.utils.ProxyAuthenticator;
+import com.sonar.it.scanner.msbuild.utils.TestUtils;
+import com.sonar.orchestrator.build.BuildResult;
+import com.sonar.orchestrator.locator.FileLocation;
+import com.sonar.orchestrator.util.NetworkUtils;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.nio.file.Path;
+import java.util.List;
+import java.util.concurrent.ConcurrentLinkedDeque;
+import org.eclipse.jetty.client.api.Request;
+import org.eclipse.jetty.proxy.ProxyServlet;
+import org.eclipse.jetty.security.ConstraintMapping;
+import org.eclipse.jetty.security.ConstraintSecurityHandler;
+import org.eclipse.jetty.security.HashLoginService;
+import org.eclipse.jetty.security.SecurityHandler;
+import org.eclipse.jetty.security.UserStore;
+import org.eclipse.jetty.server.Handler;
+import org.eclipse.jetty.server.HttpConfiguration;
+import org.eclipse.jetty.server.HttpConnectionFactory;
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.server.ServerConnector;
+import org.eclipse.jetty.server.handler.DefaultHandler;
+import org.eclipse.jetty.server.handler.HandlerCollection;
+import org.eclipse.jetty.servlet.ServletContextHandler;
+import org.eclipse.jetty.servlet.ServletHandler;
+import org.eclipse.jetty.util.security.Constraint;
+import org.eclipse.jetty.util.security.Credential;
+import org.eclipse.jetty.util.thread.QueuedThreadPool;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.io.TempDir;
+import org.sonarqube.ws.Issues.Issue;
+
+import static com.sonar.it.scanner.msbuild.sonarqube.Tests.ORCHESTRATOR;
+import static org.assertj.core.api.Assertions.assertThat;
+
+@ExtendWith(Tests.class)
+class ProxyTest {
+  private static final String PROXY_USER = "scott";
+  private static final String PROXY_PASSWORD = "tiger";
+  private static Server server;
+  private static int httpProxyPort;
+
+  private static final ConcurrentLinkedDeque<String> seenByProxy = new ConcurrentLinkedDeque<>();
+
+  @TempDir
+  public Path basePath;
+
+  @BeforeEach
+  public void setUp() {
+    seenByProxy.clear();
+  }
+
+  @AfterEach
+  public void stopProxy() throws Exception {
+    if (server != null && server.isStarted()) {
+      server.stop();
+    }
+  }
+
+  @Test
+  void proxyAuth() throws Exception {
+    startProxy(true);
+    String projectKey = "testSampleWithProxyAuth";
+    ORCHESTRATOR.getServer().restoreProfile(FileLocation.of("projects/ProjectUnderTest/TestQualityProfile.xml"));
+    ORCHESTRATOR.getServer().provisionProject(projectKey, "sample");
+    ORCHESTRATOR.getServer().associateProjectToQualityProfile(projectKey, "cs", "ProfileForTest");
+
+    Path projectDir = TestUtils.projectDir(basePath, "ProjectUnderTest");
+    String token = TestUtils.getNewToken(ORCHESTRATOR);
+    TestUtils.newScannerBegin(ORCHESTRATOR, projectKey, projectDir, token).execute(ORCHESTRATOR);
+    TestUtils.buildMSBuild(ORCHESTRATOR, projectDir);
+    BuildResult result = TestUtils.newScannerEnd(ORCHESTRATOR, projectDir, token)
+      .setEnvironmentVariable("SONAR_SCANNER_OPTS", "-Dhttp.nonProxyHosts= -Dhttp.proxyHost=localhost -Dhttp.proxyPort=" + httpProxyPort)
+      .execute(ORCHESTRATOR);
+
+    assertThat(result.getLastStatus()).isNotZero();
+    assertThat(result.getLogs()).contains("407");
+    assertThat(seenByProxy).isEmpty();
+
+    TestUtils.newScannerEnd(ORCHESTRATOR, projectDir, token)
+      .setEnvironmentVariable("SONAR_SCANNER_OPTS",
+        "-Dhttp.nonProxyHosts= -Dhttp.proxyHost=localhost -Dhttp.proxyPort=" + httpProxyPort + " -Dhttp.proxyUser=" + PROXY_USER + " -Dhttp.proxyPassword=" + PROXY_PASSWORD)
+      .execute(ORCHESTRATOR);
+
+    TestUtils.dumpComponentList(ORCHESTRATOR, projectKey);
+    TestUtils.dumpProjectIssues(ORCHESTRATOR, projectKey);
+
+    List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, projectKey);
+    // 1 * csharpsquid:S1134 (line 34)
+    assertThat(issues).hasSize(1);
+    assertLineCountForProjectUnderTest(projectKey);
+
+    assertThat(seenByProxy).isNotEmpty();
+  }
+
+  private void assertLineCountForProjectUnderTest(String projectKey) {
+    var fileKey = projectKey + ":ProjectUnderTest/Foo.cs";
+    assertThat(TestUtils.getMeasureAsInteger(projectKey, "ncloc", ORCHESTRATOR)).isEqualTo(25);
+    assertThat(TestUtils.getMeasureAsInteger(fileKey, "ncloc", ORCHESTRATOR)).isEqualTo(25);
+    assertThat(TestUtils.getMeasureAsInteger(fileKey, "lines", ORCHESTRATOR)).isEqualTo(52);
+  }
+
+  private static void startProxy(boolean needProxyAuth) throws Exception {
+    httpProxyPort = NetworkUtils.getNextAvailablePort(NetworkUtils.getLocalhost());
+
+    // Setup Threadpool
+    QueuedThreadPool threadPool = new QueuedThreadPool();
+    threadPool.setMaxThreads(500);
+
+    server = new Server(threadPool);
+
+    // HTTP Configuration
+    HttpConfiguration httpConfig = new HttpConfiguration();
+    httpConfig.setSecureScheme("https");
+    httpConfig.setSendServerVersion(true);
+    httpConfig.setSendDateHeader(false);
+
+    // Handler Structure
+    HandlerCollection handlers = new HandlerCollection();
+    handlers.setHandlers(new Handler[]{proxyHandler(needProxyAuth), new DefaultHandler()});
+    server.setHandler(handlers);
+
+    ServerConnector http = new ServerConnector(server, new HttpConnectionFactory(httpConfig));
+    http.setPort(httpProxyPort);
+    server.addConnector(http);
+
+    server.start();
+  }
+
+  private static ServletContextHandler proxyHandler(boolean needProxyAuth) {
+    ServletContextHandler contextHandler = new ServletContextHandler();
+    if (needProxyAuth) {
+      contextHandler.setSecurityHandler(basicAuth("Private!"));
+    }
+    contextHandler.setServletHandler(newServletHandler());
+    return contextHandler;
+  }
+
+  private static SecurityHandler basicAuth(String realm) {
+    HashLoginService l = new HashLoginService();
+
+    UserStore userStore = new UserStore();
+    userStore.addUser(ProxyTest.PROXY_USER, Credential.getCredential(ProxyTest.PROXY_PASSWORD), new String[]{"user"});
+
+    l.setUserStore(userStore);
+    l.setName(realm);
+
+    Constraint constraint = new Constraint();
+    constraint.setName(Constraint.__BASIC_AUTH);
+    constraint.setRoles(new String[]{"user"});
+    constraint.setAuthenticate(true);
+
+    ConstraintMapping cm = new ConstraintMapping();
+    cm.setConstraint(constraint);
+    cm.setPathSpec("/*");
+
+    ConstraintSecurityHandler csh = new ConstraintSecurityHandler();
+    csh.setAuthenticator(new ProxyAuthenticator());
+    csh.setRealmName("myrealm");
+    csh.addConstraintMapping(cm);
+    csh.setLoginService(l);
+
+    return csh;
+
+  }
+
+  private static ServletHandler newServletHandler() {
+    ServletHandler handler = new ServletHandler();
+    handler.addServletWithMapping(MyProxyServlet.class, "/*");
+    return handler;
+  }
+
+  public static class MyProxyServlet extends ProxyServlet {
+    @Override
+    protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+      seenByProxy.add(request.getRequestURI());
+      super.service(request, response);
+    }
+
+    @Override
+    protected void sendProxyRequest(HttpServletRequest clientRequest, HttpServletResponse proxyResponse, Request proxyRequest) {
+      super.sendProxyRequest(clientRequest, proxyResponse, proxyRequest);
+    }
+  }
+}
