+++ its/src/test/java/com/sonar/it/scanner/msbuild/sonarqube/ScannerMSBuildTest.java
@@ -21,7 +21,6 @@ package com.sonar.it.scanner.msbuild.sonarqube;
 
 import com.eclipsesource.json.Json;
 import com.sonar.it.scanner.msbuild.utils.EnvironmentVariable;
-import com.sonar.it.scanner.msbuild.utils.ProxyAuthenticator;
 import com.sonar.it.scanner.msbuild.utils.ScannerClassifier;
 import com.sonar.it.scanner.msbuild.utils.ScannerCommand;
 import com.sonar.it.scanner.msbuild.utils.TestUtils;
@@ -30,10 +29,6 @@ import com.sonar.orchestrator.http.HttpException;
 import com.sonar.orchestrator.locator.FileLocation;
 import com.sonar.orchestrator.util.Command;
 import com.sonar.orchestrator.util.CommandExecutor;
-import com.sonar.orchestrator.util.NetworkUtils;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
@@ -45,31 +40,9 @@ import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import org.apache.commons.io.FileUtils;
-import org.eclipse.jetty.client.api.Request;
-import org.eclipse.jetty.proxy.ProxyServlet;
-import org.eclipse.jetty.security.ConstraintMapping;
-import org.eclipse.jetty.security.ConstraintSecurityHandler;
-import org.eclipse.jetty.security.HashLoginService;
-import org.eclipse.jetty.security.SecurityHandler;
-import org.eclipse.jetty.security.UserStore;
-import org.eclipse.jetty.server.Handler;
-import org.eclipse.jetty.server.HttpConfiguration;
-import org.eclipse.jetty.server.HttpConnectionFactory;
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.server.ServerConnector;
-import org.eclipse.jetty.server.handler.DefaultHandler;
-import org.eclipse.jetty.server.handler.HandlerCollection;
-import org.eclipse.jetty.servlet.ServletContextHandler;
-import org.eclipse.jetty.servlet.ServletHandler;
-import org.eclipse.jetty.util.security.Constraint;
-import org.eclipse.jetty.util.security.Credential;
-import org.eclipse.jetty.util.thread.QueuedThreadPool;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
@@ -97,27 +70,9 @@ class ScannerMSBuildTest {
   private static final String SONAR_RULES_PREFIX = "csharpsquid:";
   // note that in the UI the prefix will be 'roslyn:'
   private static final String ROSLYN_RULES_PREFIX = "external_roslyn:";
-  private static final String PROXY_USER = "scott";
-  private static final String PROXY_PASSWORD = "tiger";
-  private static Server server;
-  private static int httpProxyPort;
-
-  private static final ConcurrentLinkedDeque<String> seenByProxy = new ConcurrentLinkedDeque<>();
 
   @TempDir
   public Path basePath;
-
-  @BeforeEach
-  public void setUp() {
-    seenByProxy.clear();
-  }
-
-  @AfterEach
-  public void stopProxy() throws Exception {
-    if (server != null && server.isStarted()) {
-      server.stop();
-    }
-  }
 
   @Test
   void testSample() throws Exception {
@@ -137,50 +92,9 @@ class ScannerMSBuildTest {
     List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, projectKey);
     // 1 * csharpsquid:S1134 (line 34)
     assertThat(issues).hasSize(1);
-    assertLineCountForProjectUnderTest(projectKey);
-  }
-
-  @Test
-  void testSampleWithProxyAuth() throws Exception {
-    startProxy(true);
-    String projectKey = "testSampleWithProxyAuth";
-    ORCHESTRATOR.getServer().restoreProfile(FileLocation.of("projects/ProjectUnderTest/TestQualityProfile.xml"));
-    ORCHESTRATOR.getServer().provisionProject(projectKey, "sample");
-    ORCHESTRATOR.getServer().associateProjectToQualityProfile(projectKey, "cs", "ProfileForTest");
-
-    Path projectDir = TestUtils.projectDir(basePath, "ProjectUnderTest");
-    String token = TestUtils.getNewToken(ORCHESTRATOR);
-    TestUtils.newScannerBegin(ORCHESTRATOR, projectKey, projectDir, token).execute(ORCHESTRATOR);
-    TestUtils.buildMSBuild(ORCHESTRATOR, projectDir);
-    BuildResult result = TestUtils.newScannerEnd(ORCHESTRATOR, projectDir, token)
-      .setEnvironmentVariable("SONAR_SCANNER_OPTS", "-Dhttp.nonProxyHosts= -Dhttp.proxyHost=localhost -Dhttp.proxyPort=" + httpProxyPort)
-      .execute(ORCHESTRATOR);
-
-    assertThat(result.getLastStatus()).isNotZero();
-    assertThat(result.getLogs()).contains("407");
-    assertThat(seenByProxy).isEmpty();
-
-    TestUtils.newScannerEnd(ORCHESTRATOR, projectDir, token)
-      .setEnvironmentVariable("SONAR_SCANNER_OPTS",
-        "-Dhttp.nonProxyHosts= -Dhttp.proxyHost=localhost -Dhttp.proxyPort=" + httpProxyPort + " -Dhttp.proxyUser=" + PROXY_USER + " -Dhttp.proxyPassword=" + PROXY_PASSWORD)
-      .execute(ORCHESTRATOR);
-
-    TestUtils.dumpComponentList(ORCHESTRATOR, projectKey);
-    TestUtils.dumpProjectIssues(ORCHESTRATOR, projectKey);
-
-    List<Issue> issues = TestUtils.projectIssues(ORCHESTRATOR, projectKey);
-    // 1 * csharpsquid:S1134 (line 34)
-    assertThat(issues).hasSize(1);
-    assertLineCountForProjectUnderTest(projectKey);
-
-    assertThat(seenByProxy).isNotEmpty();
-  }
-
-  private void assertLineCountForProjectUnderTest(String projectKey) {
-    var fileKey = projectKey + ":ProjectUnderTest/Foo.cs";
     assertThat(TestUtils.getMeasureAsInteger(projectKey, "ncloc", ORCHESTRATOR)).isEqualTo(25);
-    assertThat(TestUtils.getMeasureAsInteger(fileKey, "ncloc", ORCHESTRATOR)).isEqualTo(25);
-    assertThat(TestUtils.getMeasureAsInteger(fileKey, "lines", ORCHESTRATOR)).isEqualTo(52);
+    assertThat(TestUtils.getMeasureAsInteger(projectKey + ":ProjectUnderTest/Foo.cs", "ncloc", ORCHESTRATOR)).isEqualTo(25);
+    assertThat(TestUtils.getMeasureAsInteger(projectKey + ":ProjectUnderTest/Foo.cs", "lines", ORCHESTRATOR)).isEqualTo(52);
   }
 
   @Test
@@ -1047,93 +961,10 @@ class ScannerMSBuildTest {
     return TestUtils.newWsClient(ORCHESTRATOR);
   }
 
-  private static void startProxy(boolean needProxyAuth) throws Exception {
-    httpProxyPort = NetworkUtils.getNextAvailablePort(NetworkUtils.getLocalhost());
-
-    // Setup Threadpool
-    QueuedThreadPool threadPool = new QueuedThreadPool();
-    threadPool.setMaxThreads(500);
-
-    server = new Server(threadPool);
-
-    // HTTP Configuration
-    HttpConfiguration httpConfig = new HttpConfiguration();
-    httpConfig.setSecureScheme("https");
-    httpConfig.setSendServerVersion(true);
-    httpConfig.setSendDateHeader(false);
-
-    // Handler Structure
-    HandlerCollection handlers = new HandlerCollection();
-    handlers.setHandlers(new Handler[]{proxyHandler(needProxyAuth), new DefaultHandler()});
-    server.setHandler(handlers);
-
-    ServerConnector http = new ServerConnector(server, new HttpConnectionFactory(httpConfig));
-    http.setPort(httpProxyPort);
-    server.addConnector(http);
-
-    server.start();
-  }
-
-  private static ServletContextHandler proxyHandler(boolean needProxyAuth) {
-    ServletContextHandler contextHandler = new ServletContextHandler();
-    if (needProxyAuth) {
-      contextHandler.setSecurityHandler(basicAuth("Private!"));
-    }
-    contextHandler.setServletHandler(newServletHandler());
-    return contextHandler;
-  }
-
-  private static SecurityHandler basicAuth(String realm) {
-    HashLoginService l = new HashLoginService();
-
-    UserStore userStore = new UserStore();
-    userStore.addUser(ScannerMSBuildTest.PROXY_USER, Credential.getCredential(ScannerMSBuildTest.PROXY_PASSWORD), new String[]{"user"});
-
-    l.setUserStore(userStore);
-    l.setName(realm);
-
-    Constraint constraint = new Constraint();
-    constraint.setName(Constraint.__BASIC_AUTH);
-    constraint.setRoles(new String[]{"user"});
-    constraint.setAuthenticate(true);
-
-    ConstraintMapping cm = new ConstraintMapping();
-    cm.setConstraint(constraint);
-    cm.setPathSpec("/*");
-
-    ConstraintSecurityHandler csh = new ConstraintSecurityHandler();
-    csh.setAuthenticator(new ProxyAuthenticator());
-    csh.setRealmName("myrealm");
-    csh.addConstraintMapping(cm);
-    csh.setLoginService(l);
-
-    return csh;
-
-  }
-
-  private static ServletHandler newServletHandler() {
-    ServletHandler handler = new ServletHandler();
-    handler.addServletWithMapping(MyProxyServlet.class, "/*");
-    return handler;
-  }
-
   private List<Issue> filter(List<Issue> issues, String ruleIdPrefix) {
     return issues
       .stream()
       .filter(x -> x.getRule().startsWith(ruleIdPrefix))
       .collect(Collectors.toList());
-  }
-
-  public static class MyProxyServlet extends ProxyServlet {
-    @Override
-    protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-      seenByProxy.add(request.getRequestURI());
-      super.service(request, response);
-    }
-
-    @Override
-    protected void sendProxyRequest(HttpServletRequest clientRequest, HttpServletResponse proxyResponse, Request proxyRequest) {
-      super.sendProxyRequest(clientRequest, proxyResponse, proxyRequest);
-    }
   }
 }
